+++ make_release.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+rm -rf build dist
+pynsist installer.cfg
+python setup.py bdist_wheel sdist
+twine upload dist/*