+++ uci_cbp_demo/__init__.py
@@ -2,4 +2,6 @@
 
 __author__ = """UCI cBP demo"""
 __email__ = 'taoyil@uci.edu'
-__version__ = '0.1.0'
+__version__ = '0.2.4'
+
+from .gui import gui_noargs