+++ HISTORY.rst
@@ -10,3 +10,12 @@ History
 0.1.2 (2020-02-17)
 ------------------
 * Fix typo and restrict support to Python 3.8
+
+
+0.1.3 (2020-02-17)
+------------------
+* Add device scanner and MacOSX support
+
+0.2 (2020-02-17)
+------------------
+* Support 2 channel operation and new board